Add WKServiceEnvironment, add methods for creating donate API URLs
+ plus rename URL extension file
