+++ WKData/Sources/WKData/Models/WKServiceEnvironment.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum WKServiceEnvironment {
+    case production
+    case staging
+}
